requirements.txt
streamlit
tensorflow>=2.15.0
numpy
Pillow
scikit-learn